GROOVY-9649: Make createRange aware of left side exclusivity
Currently empty ranges are created only when to == from. This is
because range indexing in lists behaves differently: while 0<..<-1
would normally be an empty range, we can index into lists with it and
expect some output, since negative indices count from the end of the
list towards its beginning. Note that while this is true for IntRanges,
object ranges like 'a'<..<'b' become EmptyRanges.